pyproject.toml
# AI Admin Hub - pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ai-admin-hub"
version = "0.1.0"
description = "AI-powered admin hub for n8n workflows with intelligent diagnostics and automation"
authors = ["Roland <your-email@example.com>"]
readme = "README.md"
packages = [{include = "ai_admin_hub", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.9.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
rich = "^13.7.0"
httpx = "^0.25.0"
gitpython = "^3.1.40"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
pylint = "^3.0.0"
mypy = "^1.7.0"
pre-commit = "^3.6.0"

[tool.poetry.scripts]
ai-admin = "ai_admin_hub.cli:app"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pylint.messages_control]
max-line-length = 88
disable = ["C0111", "R0903"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=ai_admin_hub --cov-report=term-missing"
testpaths = ["tests"]